feat(website): reimplement emote page
