+++ website/src/routes/emotes/[id]/+layout.ts
@@ -1,0 +1,7 @@
+import type { LayoutLoadEvent } from "./$types";
+
+export function load({ params }: LayoutLoadEvent) {
+    return {
+        id: params.id,
+    };
+}
