+++ website/src/routes/emotes/[id]/+page.svelte
@@ -2,22 +2,29 @@
 	import Button from "$/components/button.svelte";
 	import ChannelPreview from "$/components/channel-preview.svelte";
 	import HideOn from "$/components/hide-on.svelte";
+	import EmoteTabs from "$/components/layout/emote-tabs.svelte";
 	import SearchBar from "$/components/search-bar.svelte";
 	import { CaretLeft, CaretRight } from "phosphor-svelte";
+	import type { LayoutData } from "./$types";
+
+	export let data: LayoutData;
 </script>
 
 <div class="navigation">
-	<div class="buttons">
-		<Button disabled style="color: var(--text-light)">
-			<CaretLeft slot="icon" />
-		</Button>
-		<Button>
-			<CaretRight slot="icon" />
-		</Button>
+	<EmoteTabs id={data.id} />
+	<div class="inputs">
+		<div class="buttons">
+			<Button disabled>
+				<CaretLeft slot="icon" />
+			</Button>
+			<Button>
+				<CaretRight slot="icon" />
+			</Button>
+		</div>
+		<HideOn mobile>
+			<SearchBar />
+		</HideOn>
 	</div>
-	<HideOn mobile>
-		<SearchBar />
-	</HideOn>
 </div>
 <div class="channels">
 	{#each Array(20) as _, i}
@@ -27,16 +34,20 @@
 
 <style lang="scss">
 	.navigation {
-		position: absolute;
-		top: 1rem;
-		right: 1rem;
-
 		display: flex;
 		align-items: center;
+		justify-content: space-between;
 		gap: 0.5rem;
+
+		.inputs {
+			display: flex;
+			align-items: center;
+			gap: 0.5rem;
+		}
 
 		.buttons {
 			display: flex;
+			align-items: center;
 		}
 	}
