+++ website/src/routes/emotes/[id]/mod-comments/+page.svelte
@@ -1,4 +1,19 @@
-<div class="mod-comments">
-	<div class="comments"></div>
-	<div class="add-comment"></div>
+<script lang="ts">
+	import EmoteTabs from "$/components/layout/emote-tabs.svelte";
+	import type { LayoutData } from "../$types";
+
+	export let data: LayoutData;
+</script>
+
+<div class="navigation">
+    <EmoteTabs id={data.id} />
 </div>
+
+<style lang="scss">
+    .navigation {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        gap: 0.5rem;
+    }
+</style>
