+++ website/src/routes/user/[username]/+page.svelte
@@ -1,1 +1,5 @@
-active
+<script lang="ts">
+	import EmotePreview from "$/components/emote-preview.svelte";
+</script>
+
+<EmotePreview />
