+++ website/src/routes/user/[username]/uploaded/+page.svelte
@@ -1,1 +1,5 @@
-uploaded
+<script lang="ts">
+	import EmotePreview from "$/components/emote-preview.svelte";
+</script>
+
+<EmotePreview name="uploadedEmote1" />
